% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_measures.R
\name{construct_measures}
\alias{construct_measures}
\title{Construct Network Measures}
\usage{
construct_measures(egor.data, include.optional = FALSE)
}
\arguments{
\item{egor.data}{An egor object.}

\item{include.optional}{Logical indicating whether to construct more obscure measures: transitivity, diameter, eff.size, brokerage, shortest.path. Defaults to FALSE.}
}
\description{
Adds commonly used network data measures onto each subnetwork level of
the input egor object.
}
