% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_vars.R
\name{define_vars}
\alias{define_vars}
\title{Define New Variables for an Egor Object}
\usage{
define_vars(
  egor.data,
  ego.formula = list(""),
  alter.formula = list(""),
  aatie.formula = list("")
)
}
\arguments{
\item{egor.data}{An egor object.}

\item{ego.formula}{An optional expression to be executed within a dplyr::mutate to define new ego-level variables.}

\item{alter.formula}{An optional expression to be executed within a dplyr::mutate to define new alter-level variables.}

\item{aatie.formula}{An optional expression to be executed within a dplyr::mutate to define new aatie-level variables.}
}
\description{
Provides a single function for defining new variables at the ego-, alter-,
and aatie-levels of an egor object.
}
