% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropIDs.R
\name{dropIDs}
\alias{dropIDs}
\title{Drop Egos or Alters from all levels of an Egor object}
\usage{
dropIDs(egor.data, ego.drops, alter.drops)
}
\arguments{
\item{egor.data}{An egor object.}

\item{ego.drops}{A vector of egos specified by .egoID for which all data should be dropped}

\item{alter.drops}{A vector of alters specified with the format ".egoID|.altID" for which all data should be dropped}
}
\description{
Provides a function for removing data for specified egos, alters of an egor object
}
