% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary_vars.R
\name{make_summary_vars}
\alias{make_summary_vars}
\title{Create Summary Variables for an Egor Object}
\usage{
make_summary_vars(
  egor.data,
  codebook.address,
  additional.multselect.vars = NULL,
  additional.multselect.vars.labels = list(""),
  prop.denom = "struct.alters",
  additional.numeric.vars = NULL,
  output.csv.address = ""
)
}
\arguments{
\item{egor.data}{An egor object.}

\item{codebook.address}{The address to the corresponding EgoWeb codebook.}

\item{additional.multselect.vars}{An optional vector of strings listing variables not in the codebook that should be processed as multiple selection variables.}

\item{additional.multselect.vars.labels}{A list providing corresponding labels to be used with the additional.multselect.vars.}

\item{prop.denom}{A string specifying how denominators for ego-level proportions should be calculated, where 'struct.alters' refers to the total alter count for each ego.}

\item{additional.numeric.vars}{An optional vector of strings listing variables not in the codebook that should be processed as numeric.}

\item{output.csv.address}{An optional address to which a summary of the newly constructed variables should be written.}
}
\description{
Uses an EgoWeb codebook to add statistical summary variables onto the
input egor object -- dummy variables for multiple selection question values,
recoded categorical variables using the codebook labels,
ego-level totals, ego-level means, and ego-level proportions.
}
