% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_egoweb.R
\name{process_egoweb}
\alias{process_egoweb}
\title{Process EgoWeb export data}
\usage{
process_egoweb(
  alter.csv = "",
  edges.csv = "",
  egos.csv = NULL,
  ID.vars = list(ego = "EgoID", alter = "Alter.Number", source = "Alter.1.Number", target
    = "Alter.2.Number"),
  alter.drops = "",
  edges.drops = "",
  egos.drops = "",
  return.egor = TRUE,
  ego.vars = NULL,
  ...
)
}
\arguments{
\item{alter.csv}{A .csv file containing alter-level data.}

\item{edges.csv}{A .csv file containing alter-alter data.}

\item{egos.csv}{A .csv file containing ego-level data, optional.}

\item{ID.vars}{A vector specifying the identifier variable for each of: ego, alter, source of an alter-alter link, and target of an alter-alter link.}

\item{alter.drops}{An optional string specifying a condition to identify records that should be dropped from the alter-level data.}

\item{edges.drops}{An optional string specifying a condition to identify records that should be dropped from the alter-alter data.}

\item{egos.drops}{An optional string specifying a condition to identify records that should be dropped from the ego-level data.}

\item{return.egor}{A logical value to indicate whether the EgoWeb data should be restructured and returned as an egor class object. Defaults to TRUE.}

\item{ego.vars}{An optional vector listing variables from the ego-level dataset that should be merged onto the alter-level data.}
}
\description{
Takes .csv data exported from EgoWeb, removes certain undesirable
observations, and can optionally return the data structured as
an egor-class object to cooperate with the egor package.
}
