% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_subnetworks.R
\name{split_subnetworks}
\alias{split_subnetworks}
\title{Split Egor Data Into Subnetworks}
\usage{
split_subnetworks(
  egor.data,
  alter.formula.binary = list(""),
  aatie.formula.binary = list("")
)
}
\arguments{
\item{egor.data}{An egor object.}

\item{alter.formula.binary}{An expression to be executed within a dplyr::mutate to define a new binary variable to distinguish separate alter-level networks.}

\item{aatie.formula.binary}{An expression to be executed within a dplyr::mutate to define a new binary variable to distinguish separate aatie-level networks.}
}
\description{
Uses provided expressions to define indicator variables for different
subnetwork levels and add another level to the egor input object
that specifies each subnetwork
}
