% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_alter.R
\name{table1_alter}
\alias{table1_alter}
\title{Alter Summary Table}
\usage{
table1_alter(egor.data, caption = NULL)
}
\arguments{
\item{egor.data}{An egor object.}

\item{caption}{An optional string to put at the top of the table}
}
\description{
Produces a formatted table of alter-level summary variables created by
make_summary_vars or construct_measures
}
