% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_ego.R
\name{table1_ego}
\alias{table1_ego}
\title{Ego Summary Table}
\usage{
table1_ego(egor.data, caption = NULL)
}
\arguments{
\item{egor.data}{An egor object.}

\item{caption}{An optional string to put at the top of the table}
}
\description{
Produces a formatted table of ego-level summary variables created by
make_summary_vars or construct_measures
}
