% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_struct.R
\name{table1_struct}
\alias{table1_struct}
\title{Structural Measure Summary Table}
\usage{
table1_struct(egor.data, caption = NULL)
}
\arguments{
\item{egor.data}{An egor object.}

\item{caption}{An optional string to put at the top of the table}
}
\description{
Produces a formatted table of structural measures created by construct_measures
}
