% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_egor.R
\name{visualize_egor}
\alias{visualize_egor}
\title{Visualize an Egor Object with R Shiny}
\usage{
visualize_egor(object = NULL, shiny_opts = list(launch.browser = TRUE))
}
\arguments{
\item{object}{An egor-class object.}

\item{shiny_opts}{A list of arguments to be passed to shiny::shinyApp's options argument.}
}
\description{
A custom version of egor::egor_vis_app with minor modifications
}
