% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_measures.R
\name{construct_measures}
\alias{construct_measures}
\title{Construct Network Measures}
\usage{
construct_measures(egor.data)
}
\arguments{
\item{egor.data}{An egor object.}
}
\description{
Adds commonly used network data measures onto each subnetwork level of
the input egor object.
}
